Order registration WORKS!

+++ src/main/java/com/mas/project/mas/Service/OrderService.java
@@ -8,7 +8,7 @@ import com.mas.project.mas.Entity.Order;
  * Created by Janusz on 07.06.2017.
  */
 public interface OrderService {
-    Order findByOrderNumber(Long orderNumber);
+    Order findByOrderNumber(Integer orderNumber);
     Order saveOrder(OrderDTO order);
     Order updateOrder(Order order);
     Order setOrderStatus(Order order, Status status);
